+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 
 ## Upcoming
-
+- **UPDATE**: Support for LM75AD Sensor [#275](https://github.com/reloxx13/TasmoAdmin/issues/275)
 
 ### v1.6.1-beta2
 -FIX: login box position fix